refactor: 디자인 시스템 컴포넌트 패키지 구조 변경 (modal -> feedback)
`PrezelDialog` 및 `PrezelSnackbar` 관련 컴포넌트의 위치를 `modal` 패키지에서 `feedback` 패키지로 이동하고 관련 임포트 경로를 수정하였습니다.

* **refactor: Dialog 및 Snackbar 컴포넌트 패키지 이동**
    * `component.modal.dialog` -> `component.feedback.dialog`
    * `component.modal.snackbar` -> `component.feedback.snackbar`
    * 이동된 파일: `PrezelDialog.kt`, `PrezelDialogScope.kt`, `PrezelSnackbar.kt`, `SnackbarHost.kt`

* **refactor: 패키지 변경에 따른 의존성 업데이트**
    * `PrezelNavigationBar`, `ProfileScreen`, `HistoryScreen`, `HomeScreen`, `LoginScreen`, `DoubleBackToExitHandler` 등에서 변경된 패키지 경로로 `import` 문 수정
