+++ Prezel/core/designsystem/src/main/java/com/team/prezel/core/designsystem/component/feedback/dialog/PrezelDialogScope.kt
@@ -1,4 +1,4 @@
-package com.team.prezel.core.designsystem.component.modal.dialog
+package com.team.prezel.core.designsystem.component.feedback.dialog
 
 import androidx.compose.foundation.layout.LayoutScopeMarker
 import androidx.compose.foundation.layout.padding
